Extract progress bar owner
